Set up separate files for classes/functions in manager/utils.py; moved EmailMessage to its own file
